Organize/cleanup examples and ruleaction
