--- ruleaction/flogo.json
@@ -1,186 +1,0 @@
-{
-   "name":"mysixthapp",
-   "type":"flogo:app",
-   "version":"0.0.1",
-   "description":"My flogo application description",
-   "triggers":[
-      {
-         "id":"my_rest_trigger",
-         "ref":"github.com/TIBCOSoftware/flogo-contrib/trigger/rest",
-         "settings":{
-            "port":"9233"
-         },
-         "handlers":[
-            {
-               "name":"customerevent",
-               "action":{
-                  "id":"ruleaction"
-               },
-               "settings":{
-                  "method":"GET",
-                  "path":"/cust"
-               }
-            },{
-               "name":"debitevent",
-               "action":{
-                  "id":"ruleaction"
-               },
-               "settings":{
-                  "method":"GET",
-                  "path":"/debit"
-               }
-            },{
-               "name":"packageevent",
-               "action":{
-                  "id":"ruleaction"
-               },
-               "settings":{
-                  "method":"GET",
-                  "path":"/pkg"
-               }
-            },
-            {
-               "name":"scanevent",
-               "action":{
-                  "id":"ruleaction"
-               },
-               "settings":{
-                  "method":"GET",
-                  "path":"/scan"
-               }
-            }
-         ]
-      }
-   ],
-   "actions":[
-      {
-         "id":"ruleaction",
-         "name":"Rule action",
-         "ref":"github.com/project-flogo/rules/ruleaction",
-         "data":{
-            "Tds":[
-              {
-                 "name":"customerevent",
-                 "ttl":-1,
-                 "properties":[
-                   {     
-                       "name":"name",
-                       "type":"string"
-                    },
-                    {
-                       "name":"status",
-                       "type":"string"
-                    },
-                    {
-                       "name":"balance",
-                       "type":"string"
-                    },
-                    {
-                       "name":"addr",
-                       "type":"string"
-                    }
-                   ]
-              },
-              {
-                 "name":"debitevent",
-                 "ttl":0,
-                 "properties":[
-                    {
-                       "name":"name",
-                       "type":"string"
-                    },
-                    {
-                       "name":"debit",
-                       "type":"string"
-                    }
-                   ]
-              },
-              {
-                 "name":"packageevent",
-                 "ttl":0,
-                 "properties":[
-                    {
-                       "name":"packageid",
-                       "type":"string"
-                    },
-                    {
-                       "name":"next",
-                       "type":"string"
-                    },
-                    {
-                       "name":"status",
-                       "type":"string"
-                    },
-                    {
-                       "name":"isnew",
-                       "type":"string"
-                    }
-                   ]
-              },
-              {
-                 "name":"package",
-                 "ttl":-1,
-                 "properties":[
-                    {
-                       "name":"packageid",
-                       "type":"string"
-                    },
-                    {
-                       "name":"curr",
-                       "type":"string"
-                    },
-                    {
-                       "name":"next",
-                       "type":"string"
-                    },
-                    {
-                       "name":"status",
-                       "type":"string"
-                    },
-                    {
-                       "name":"isnew",
-                       "type":"string"
-                    }
-                   ]
-              },
-              {
-                 "name":"scanevent",
-                 "ttl":0,
-                 "properties":[
-                    {
-                       "name":"packageid",
-                       "type":"string"
-                    },
-                    {
-                       "name":"curr",
-                       "type":"string"
-                    },
-                    {
-                       "name":"next",
-                       "type":"string"
-                    },
-                    {
-                       "name":"eta",
-                       "type":"string"
-                    }
-                   ]
-              },
-              {
-                 "name":"scantimeout",
-                 "ttl":0,
-                 "properties":[
-                    {
-                       "name":"packageid",
-                       "type":"string"
-                    },
-                    {
-                       "name":"next",
-                       "type":"string"
-                    }
-                   ]
-              }
-            ]
-         }
-      }
-   ]
-}
